Move methods to Mod repository
